[Fix] #39 - merge 후 confilct 해결

+++ src/main/java/org/moonshot/server/domain/keyresult/service/KeyResultService.java
@@ -18,8 +18,11 @@ import org.moonshot.server.domain.log.service.LogService;
 import org.moonshot.server.domain.objective.exception.ObjectiveNotFoundException;
 import org.moonshot.server.domain.objective.model.Objective;
 import org.moonshot.server.domain.objective.repository.ObjectiveRepository;
+import org.moonshot.server.domain.task.dto.request.TaskCreateRequestDto;
 import org.moonshot.server.domain.task.model.Task;
 import org.moonshot.server.domain.task.repository.TaskRepository;
+import org.moonshot.server.domain.task.service.TaskService;
+import org.moonshot.server.domain.user.service.UserService;
 import org.moonshot.server.global.common.model.Period;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -34,6 +37,8 @@ public class KeyResultService {
     private final ObjectiveRepository objectiveRepository;
     private final KeyResultRepository keyResultRepository;
     private final TaskRepository taskRepository;
+    private final TaskService taskService;
+    private final UserService userService;
     private final LogService logService;
 
     //TODO
@@ -60,16 +65,20 @@ public class KeyResultService {
                         .idx(task.idx())
                         .keyResult(keyResult)
                         .build()).toList());
+                for (TaskCreateRequestDto taskDto : dto.taskList()) {
+                    taskService.saveTask(keyResult, taskDto);
+                }
             }
         }
     }
 
     @Transactional
-    public void createKeyResult(KeyResultCreateRequestDto request) {
-        Objective objective = objectiveRepository.findById(request.objectiveId())
+    public void createKeyResult(KeyResultCreateRequestDto request, Long userId) {
+        Objective objective = objectiveRepository.findObjectiveAndUserById(request.objectiveId())
                 .orElseThrow(ObjectiveNotFoundException::new);
-        List<KeyResult> krList = keyResultRepository.findAllByObjective(objective);
+        userService.validateUserAuthorization(objective.getUser(), userId);
 
+        List<KeyResult> krList = keyResultRepository.findAllByObjective(objective);
         if (krList.size() >= ACTIVE_KEY_RESULT_NUMBER) {
             throw new KeyResultNumberExceededException();
         }
@@ -93,9 +102,11 @@ public class KeyResultService {
     }
 
     @Transactional
-    public void deleteKeyResult(Long keyResultId) {
-        KeyResult keyResult = keyResultRepository.findById(keyResultId)
+    public void deleteKeyResult(Long keyResultId, Long userId) {
+        KeyResult keyResult = keyResultRepository.findKeyResultAndObjective(keyResultId)
                 .orElseThrow(KeyResultNotFoundException::new);
+        userService.validateUserAuthorization(keyResult.getObjective().getUser(), userId);
+
         taskRepository.deleteAllInBatch(taskRepository.findAllByKeyResult(keyResult));
         keyResultRepository.delete(keyResult);
     }
@@ -111,9 +122,10 @@ public class KeyResultService {
     }
 
     @Transactional
-    public void modifyKeyResult(KeyResultModifyRequestDto request) {
-        KeyResult keyResult = keyResultRepository.findById(request.keyResultId())
+    public void modifyKeyResult(KeyResultModifyRequestDto request, Long userId) {
+        KeyResult keyResult = keyResultRepository.findKeyResultAndObjective(request.keyResultId())
                 .orElseThrow(KeyResultNotFoundException::new);
+        userService.validateUserAuthorization(keyResult.getObjective().getUser(), userId);
 
         if (request.title() != null) {
             keyResult.modifyTitle(request.title());
